refactor: Change the mainMenuOptions() method to the Menus class, where all the menus will be located.
